Reorganizing scripts
