+++ Assets/Scripts/Interactables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f4f3fd7481e862849afafdded60b1182
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
