+++ Assets/Scripts/Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0145ac51313fa5848bc090f32bb8c3b0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
